feat: extract ScopesView from SettingsView (#1203)
Mirrors the LocalRunnersView extraction (#1202). Moves all remote scope
UI and logic from SettingsView into a new ScopesView, accessible via a
"Manage scopes" navigation row.

Changes:
- Add Sources/RunnerBar/Views/Settings/ScopesView.swift
  Owns @StateObject scopeStore, showAddScopeSheet + selectedScopeEntry
  @State, remoteScopesSectionHeader, remoteScopesSection, scopeRow,
  AddScopeSheet sheet, and ScopeEditSheet sheet.

- Update Sources/RunnerBar/Views/Settings/SettingsView.swift
  Remove @StateObject scopeStore, showAddScopeSheet, selectedScopeEntry,
  scopeMutateCancellable, remoteScopesSectionHeader, remoteScopesSection,
  scopeRow, both scope-related .sheet modifiers, and the
  scopeMutateCancellable assignment in onAppearAction.
  Add @State private var showScopes = false.
  body if/else now also handles showScopes -> ScopesView.
  sectionsStack replaces remoteScopesSection+Divider with
  manageScopesRow+Divider.

SwiftLint: file_header, missing_docs, sorted_imports satisfied.
Periphery: ScopesView referenced from SettingsView - not orphaned.
Tests: no model/store changes; RunnerBarCoreTests/RunnerBarUITests unaffected.

Closes #1203
